Tarefa2_2021li1g075

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2021/22.

Documentation

constroiMapalinhas :: [(Peca, Coordenadas)] -> Int -> Int -> Int -> [Peca]

A função constroiMapalinhas recebe uma lista de peças com as suas coordenadas, e três inteiros (numero da linha, x inicial, e o numero de x a utilizar) na qual retribui uma lista de peças.

Esta função é utilizada para construir as linhas do mapa.

Exemplo:

>>> constroiMapalinhas [(Bloco, (0, 4)), (Bloco, (1, 4)), (Bloco, (2, 4)),(Bloco, (3, 4)),(Bloco, (4, 4))] 4 0 5
[Bloco,Bloco,Bloco,Bloco,Bloco]

constroiMapacolunas :: [(Peca, Coordenadas)] -> Int -> Int -> Int -> Mapa

A função constroiMapacolunasrecebe uma lista de peças com as suas coordenadas, e três inteiros (y inicial, e o numero de x e de y a utilizar) na qual retribui uma lista de peças.

Este função é utilizada para construir as colunas do mapa

Exemplo:

>>> constroiMapacolunas [(Porta, (1, 0)), (Bloco, (2, 0)), (Bloco, (3, 0)), (Bloco, (4, 0))] 0 6 1
[[Vazio,Porta,Bloco,Bloco,Bloco,Vazio]]

constroiMapa :: [(Peca, Coordenadas)] -> Mapa

A função constroiMapa recebe uma listas de peças com coordenadas e fornece um mapa.

A lista de peças antes de ser fornecida terá de ser confirmada pela função validaPotencialMapa para ser considerado como válido e só assim ser construído.

Esta função é apenas a aplicação das funções constroiMapacolunas e constroiMapalinhas em conjunto, que irão formar o mapa.

Exemplo:

>>> constroiMapa [(Porta,(2,1)) ,(Bloco, (1,1)), (Bloco, (2,0))]
[[Vazio,Vazio,Vazio],[Vazio,Bloco,Porta],[Vazio,Vazio,Bloco]]

desconstroiMapa :: Mapa -> [(Peca, Coordenadas)]

A função desconstroiMapa recebe um mapa e fornece uma lista de peças.

Esta função é a aplicação inversa da função "constroiMapa"

Esta função é apenas a aplicação das funções desconstroiMapacolunas e desconstroiMapalinhas em conjunto, que irão formar a lista de peças.

Exemplo:

>>> desconstroiMapa [[Vazio,Vazio,Vazio],[Vazio,Bloco,Porta],[Vazio,Vazio,Bloco]]
[(Porta,(2,1)) ,(Bloco, (1,1)), (Bloco, (2,0))]

desconstroiMapacolunas :: Mapa -> Int -> [(Peca, Coordenadas)]

A função desconstroiMapacolunasrecebe um mapa e um inteiros (y da primeira lista) na qual retribui uma lista de peças e coordenadas.

Este função é utilizada para desconstruir as colunas do mapa

Exemplo:

>>> desconstroiMapacolunas [[Vazio,Porta,Bloco,Bloco,Bloco,Vazio]] 0
[(Porta,(1,0)),(Bloco,(2,0)),(Bloco,(3,0)),(Bloco,(4,0))]

desconstroiMapalinhas :: [Peca] -> Int -> Int -> [(Peca, Coordenadas)]

A função desconstroiMapalinhas recebe um mapa, e dois inteiros (y das peças e o x da primeira peça) na qual retribui uma lista de peças e coordenadas.

Esta função é utilizada para desconstruir as linhas do mapa.

Exemplo:

>>> desconstroiMapalinhas [Bloco,Bloco,Bloco,Bloco,Bloco] 4 0
[(Bloco,(0,4)),(Bloco,(1,4)),(Bloco,(2,4)),(Bloco,(3,4)),(Bloco,(4,4))]