Tarefa3_2021li1g075

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2021/22.

Documentation

convertertexto :: Jogo -> String

A função convertertexto serve transformar um mapa, e demonstrar a posição do jogador no mapa, em algo mais gráfico, transformando as peças e o jogador em símbolos.

Esta função utiliza duas funções auxiliares, que transformam o mapa e o jogador nesses simbolos, e os organizam em linhas e colunas, em forma de string.

Exemplo:

>>> convertertexto (Jogo [[Vazio,Vazio,Vazio],[Vazio,Vazio,Vazio],[Bloco,Bloco,Bloco]] (Jogador (1,1) Este False))
"   
 > 
XXX"

Orphan instances

Show Jogo

Instance details

Methods

showsPrec :: Int -> Jogo -> ShowS

show :: Jogo -> String

showList :: [Jogo] -> ShowS