Tarefa4_2021li1g075

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2021/22.

Documentation

moveJogador :: Jogo -> Movimento -> Jogo

A função moveJogador recebe um Jogo e um Movimento retribuindo assim o jogo atualizado, ou seja depois do Jogador se ter movido.

Esta função utiliza outras funções auxiliares que tanto aplicam a mudança no mapa, como aplicam o movimento ao Jogador.

correrMovimentos :: Jogo -> [Movimento] -> Jogo

A função correMovimentos aplica a função moveJogador a todos os elementos de uma lista ordenada de Movimento.

pegarCaixa :: Jogo -> Jogo

A função pegarCaixa recebe um Jogo e retribui um Jogo sem a Caixa no mapa mas sim com o Jogador.

Esta função utiliza outras funções auxiliares que removem a Caixa do mapa e alteram o Bool do Jogador, desta forma colocando a Caixa no Jogador

largarCaixa :: Jogo -> Jogo

A função largar recebe um Jogo e retribui um Jogo com a Caixa no mapa e não com o Jogador.

Esta função utiliza outras funções auxiliares que colocam a Caixa no mapa e alteram o Bool do jogador, desta forma retirando a Caixa do Jogador.

moveJogadorEsq :: Jogo -> Jogo

A função moveJogadorEsq recebe um Jogo e retribui um Jogo com o Jogador numa outra posição mais à esquerda.

Esta função utiliza outras funções auxiliares que movem o jogador para a esquerda, alterando o seu x.

moveJogadorDir :: Jogo -> Jogo

A função moveJogadorEsq recebe um Jogo e retribui um Jogo com o Jogador numa outra posição mais à direita.

Esta função utiliza outras funções auxiliares que movem o jogador para a direita, alterando o seu x.

moveJogadorDirCaixa :: Jogo -> Jogo

A função moveJogadorDirCaixa tem o mesmo objetivo da função moveJogadorDir, mas é utilizada quando esta tem a Caixa.

moveJogadorEsqCaixa :: Jogo -> Jogo

A função moveJogadorEsqCaixa tem o mesmo objetivo da função moveJogadorEsq, mas é utilizada quando esta tem a Caixa.

meterCaixa :: Jogo -> Jogo

A função meterCaixa é a função auxiliar utilizada em outras funções para colocar a Caixa no Jogo e ainda mudar o Bool do Jogador.

Esta função usa ainda funções auxiliares para colocar a Caixa do Mapa.

moveJogadorCair :: Jogo -> Jogo

A função moveJogadorCair é utilizada para aplicar o momento de queda ou descida do jogador, este movimento é aplicado mudando a posição do jogador no Mapa.

moveJogadorTrepar :: Jogo -> Jogo

A função moveJogadorTrepar aplica o movimento de subida ao Jogador através da alteração da posição do Jogador no Mapa.

moveJogadorTreparCaixa :: Jogo -> Jogo

A função moveJogadorTreparCaixa tem o mesmo objetivo da função moveJogadorTrepar mas é usada quando o Jogador carrega uma Caixa.

tirarCaixa :: Jogo -> Jogo

A função tirarCaixa é a função auxiliar utilizada em outras funções para tirar a Caixa do Jogo e ainda mudar o Bool do Jogador.

Esta função usa ainda funções auxiliares para retirar a Caixa do Mapa.